Advanced_Agentic_Chunking_RAG/requirements.txt
langchain>=0.0.267
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
nltk>=3.8.1
scikit-learn>=1.3.0
tqdm>=4.66.1
faiss-cpu>=1.7.4 